create separate package for expenses
