+++ backend/src/main/java/xyz/stasiak/cobudgetbackend/moneyentry/expense/MonthlyExpensesRepository.java
@@ -1,4 +1,4 @@
-package xyz.stasiak.cobudgetbackend.moneyentry;
+package xyz.stasiak.cobudgetbackend.moneyentry.expense;
 
 import java.util.Optional;
